backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "caldav>=2.1.2",
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt[crypto]>=5.5.1",
    "gunicorn>=23.0.0",
    "icalendar>=6.3.2",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "psycopg-binary>=3.1.4",
    "dj-database-url>=2.2.0",
    "whitenoise>=6.11.0",
    "psycopg2-binary>=2.9.11",
    "mysqlclient>=2.2.7",
    "requests>=2.32.5",
    "mysql-connector-python>=9.5.0",
]
